pyproject.toml
[tool.poetry]
name = "hft-tradingbot"
version = "0.1.0"
description = "High-Frequency Trading Bot with ML and Backtesting"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
websockets = "^12.0"

# Database
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"  # PostgreSQL
aiosqlite = "^0.19.0"  # Async SQLite

# API Clients
polygon-api-client = "^1.13.2"
alpaca-py = "^0.20.0"
requests = "^2.31.0"
aiohttp = "^3.9.1"

# Data Processing
pandas = "^2.1.4"
numpy = "^1.26.3"
ta-lib = "^0.4.28"  # Technical Analysis
python-dateutil = "^2.8.2"

# Machine Learning
tensorflow = "^2.15.0"
scikit-learn = "^1.4.0"
stable-baselines3 = "^2.2.1"
gymnasium = "^0.29.1"  # OpenAI Gym replacement

# Monitoring & Performance
psutil = "^5.9.7"
prometheus-client = "^0.19.0"

# Configuration & Security
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

# Logging
loguru = "^0.7.2"
structlog = "^24.1.0"

# Utilities
pytz = "^2023.3"
schedule = "^1.2.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"  # For testing async endpoints

# Linting & Formatting
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

# Load Testing
locust = "^2.20.0"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=backend --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
